. rtfiles.cpp, rtdump.cpp, routmain.h, rtdump.cpp, routmain.cpp: split rtfiles to rtfiles.cpp, rtdump.cpp, and rtdump.cpp can be shared with the upcoming appman support. . fuseif.h, fuseif.cpp, ifstat.cpp, ifstat.h: move IfStateToString from fuseif.cpp to ifstat.cpp so that it can be shared by other modules. blkalloc.h, regfs.cpp: add a new version of GetAttr with RFHANDLE.
