+++ fuse/fuseif.cpp
@@ -994,40 +994,6 @@ gint32 CFuseTextFile::fs_read(
     return ret;
 }
 
-const char* IfStateToString( EnumIfState dwState )
-{
-    switch( dwState )
-    {
-    case stateStopped:
-        return "Stopped";
-    case stateStarting:
-        return "Starting";
-    case stateStarted:
-        return "Started";
-    case stateConnected:
-        return "Connected";
-    case stateRecovery:
-        return "Recovery";
-    case statePaused:
-        return "Paused";
-    case stateUnknown:
-        return "Unknown";
-    case stateStopping:
-        return "Stopping";
-    case statePausing:
-        return "Pausing";
-    case stateResuming:
-        return "Resuming";
-    case stateIoDone:
-        return "IoDone";
-    case stateStartFailed:
-        return "StartFailed";
-    case stateInvalid:
-    default:
-        return "Invalid";
-    }
-}
-
 static gint32 GetFormatTime(
     time_t epoc_sec, stdstr& strTime )
 {
