+++ rpc/router/Makefile.am
@@ -1,5 +1,5 @@
 rpcrouter_PROGRAMS=rpcrouter
-rpcrouter_SOURCES=routmain.cpp routmain.h rpcrouter.1
+rpcrouter_SOURCES=routmain.cpp routmain.h rpcrouter.1 rtdump.cpp
 
 rpcrouter_LDADD=
 if ENABLE_FUSE3
@@ -29,7 +29,7 @@ OFFSET_MACRO=
 endif
 
 EXTRA_DIST=README.md README_cn.md
-rpcrouter_CPPFLAGS= -I$(include_rpcdir) -I$(include_ipcdir) -fPIC $(OFFSET_MACRO) 
+rpcrouter_CPPFLAGS= -I$(top_srcdir)/monitor/client/cpp/appmancli -I$(include_rpcdir) -I$(include_ipcdir) -fPIC $(OFFSET_MACRO) 
 
 if ENABLE_FUSE3
 include_fusedir=$(top_srcdir)/fuse
