+++ rpc/router/routmain.cpp
@@ -29,10 +29,10 @@
 
 #include <rpc.h>
 using namespace rpcf;
-#include "routmain.h"
 #include <ifhelper.h>
 #include <frmwrk.h>
 #include <rpcroute.h>
+#include "routmain.h"
 
 #ifdef FUSE3
 #include <fuseif.h>
